update testing format
